Renamed Snowshot class to Kamikazesnowball
This is necessary as the drag&drop function for the Dispenser uses the
class name, not the SupertuxObject name and 'snowshot' isn't a valid
object in SuperTux.

This change will require a resave of levels, as the sorting of
gameobjects is based on the class name as well, not the SupertuxObject
name.